feat: rank section style
